requirements.txt
requests
beautifulsoup4
tweepy
python-dotenv
schedule
google-generativeai
